requirements.txt
streamlit
google-genai
reportlab
python-docx